% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{insert_with_timestamp}
\alias{insert_with_timestamp}
\title{Insert a data.frame into mongo with a creation time stamp.}
\usage{
insert_with_timestamp(
  data,
  collection,
  ts_field = "modified",
  cr_field = "created",
  ...
)
}
\arguments{
\item{data}{A data.frame to add to a mongo collection.}

\item{collection}{A mongo colleciton object where the data will be stored.}

\item{ts_field}{Name of key for the modified time stamp.}

\item{cr_field}{Name of key for the created time stamp.}
}
\value{
The result from \code{\link{mongolite::mongo}$insert}.
}
\description{
A wrapper around \code{\link{mongolite::mongo}$insert}.  Mongoserver is
queried for current time and that time is add to the data.frame in two
columns named using the values "ts_field" (modified time/date) and "cr_field"
(created time/date).
}
