% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{j}
\alias{j}
\title{Easily make JSON from named parameters.}
\usage{
j(..., args = list(auto_unbox = TRUE, POSIXt = "mongo"))
}
\arguments{
\item{...}{any named objects to be put in the JSON string.}

\item{args}{A named list of addition arguments to be passed to
\code{\link{toJSON}}}
}
\value{
A JSON string.
}
\description{
A shortcut for making JSON.  \code{j(a=1, b=2)} will return the same as
\code{jsonlite::toJSON(list(a=1, b=2))}.  Nested lists are allowed, e.g.,
\code{j(a=1, b=list(c=2, d=3))}
}
