% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\name{unbox_dates}
\alias{unbox_dates}
\title{Tag singleton dates in a list for unboxing during conversion to JSON.}
\usage{
unbox_dates(document)
}
\arguments{
\item{document}{A list (or list-derived) object intended to be converted to
JSON.}
}
\value{
A copy of the list with singleton POSIXt values tagged for unboxing.
}
\description{
Uses \code{\link{rapply()}} to search all levels of a list or object
built atop a list for POSIXt vectors.  Any POSIXt vector with length
of 1 will be marked for unboxing.  This function is useful because
\code{auto_unbox} in \code{\link{toJSON}} does not auto_unbox
POSIXt values.
}
